refacor(react-components): Move Mock files to test area and change import statement to use ##test-utils (#5108)
* Move files

* Update vite.config.ts

* Test CropBoxDomainObject, SliceDomainObject and ClipFolder

* Revert "Test CropBoxDomainObject, SliceDomainObject and ClipFolder"

This reverts commit c7b53fc14c812c446c559dd8a2417c748121fc23.
