+++ react-components/src/architecture/base/utilities/primitives/Vector3ArrayUtils.test.ts
@@ -5,10 +5,7 @@
 import { describe, expect, test } from 'vitest';
 import { Vector3ArrayUtils } from './Vector3ArrayUtils';
 import { Box3, Vector3 } from 'three';
-import {
-  expectEqualBox3,
-  expectEqualVector3
-} from '../../../../../tests/tests-utilities/primitives/primitiveTestUtil';
+import { expectEqualBox3, expectEqualVector3 } from '#test-utils/primitives/primitiveTestUtil';
 
 describe(Vector3ArrayUtils.name, () => {
   const polygon = [
