+++ react-components/tests/tests-utilities/architecture/mock-commands/MockEnumOptionCommand.ts
@@ -2,9 +2,9 @@
  * Copyright 2024 Cognite AS
  */
 
-import { BaseOptionCommand } from '../BaseOptionCommand';
-import { RenderTargetCommand } from '../RenderTargetCommand';
-import { type TranslationInput } from '../../utilities/TranslateInput';
+import { BaseOptionCommand } from '../../../../src/architecture/base/commands/BaseOptionCommand';
+import { RenderTargetCommand } from '../../../../src/architecture/base/commands/RenderTargetCommand';
+import { type TranslationInput } from '../../../../src/architecture/base/utilities/TranslateInput';
 
 enum MockEnum {
   Red = 'Red',
