+++ react-components/tests/tests-utilities/architecture/mock-commands/MockFilterCommand.ts
@@ -3,8 +3,11 @@
  */
 
 import { Color } from 'three';
-import { BaseFilterCommand, BaseFilterItemCommand } from '../BaseFilterCommand';
-import { type TranslationInput } from '../../utilities/TranslateInput';
+import {
+  BaseFilterCommand,
+  BaseFilterItemCommand
+} from '../../../../src/architecture/base/commands/BaseFilterCommand';
+import { type TranslationInput } from '../../../../src/architecture/base/utilities/TranslateInput';
 
 export class MockFilterCommand extends BaseFilterCommand {
   // ==================================================
