+++ react-components/tests/tests-utilities/architecture/mock-commands/MockSectionCommand.ts
@@ -1,8 +1,8 @@
 /*!
  * Copyright 2024 Cognite AS
  */
-import { type TranslationInput } from '../../utilities/TranslateInput';
-import { SectionCommand } from '../SectionCommand';
+import { type TranslationInput } from '../../../../src/architecture/base/utilities/TranslateInput';
+import { SectionCommand } from '../../../../src/architecture/base/commands/SectionCommand';
 
 export class MockSectionCommand extends SectionCommand {
   public override get tooltip(): TranslationInput {
