+++ react-components/tests/tests-utilities/architecture/mock-commands/MockSliderCommand.ts
@@ -2,8 +2,8 @@
  * Copyright 2024 Cognite AS
  */
 
-import { type TranslationInput } from '../../utilities/TranslateInput';
-import { FractionSliderCommand } from '../FractionSliderCommand';
+import { type TranslationInput } from '../../../../src/architecture/base/utilities/TranslateInput';
+import { FractionSliderCommand } from '../../../../src/architecture/base/commands/FractionSliderCommand';
 
 export class MockSliderCommand extends FractionSliderCommand {
   public _value = 0.5;
