+++ react-components/tests/tests-utilities/architecture/mock-commands/MockToggleCommand.ts
@@ -2,22 +2,17 @@
  * Copyright 2024 Cognite AS
  */
 
-import { type IconName } from '../../utilities/IconName';
-import { type TranslationInput } from '../../utilities/TranslateInput';
-import { RenderTargetCommand } from '../RenderTargetCommand';
+import { type TranslationInput } from '../../../../src/architecture/base/utilities/TranslateInput';
+import { RenderTargetCommand } from '../../../../src/architecture/base/commands/RenderTargetCommand';
 
-export class MockCheckableCommand extends RenderTargetCommand {
+export class MockToggleCommand extends RenderTargetCommand {
   public value = false;
   // ==================================================
   // OVERRIDES
   // ==================================================
 
   public override get tooltip(): TranslationInput {
-    return { untranslated: 'Checkable action' };
-  }
-
-  public override get icon(): IconName {
-    return 'Snow';
+    return { untranslated: 'Boolean' };
   }
 
   public override get isToggle(): boolean {
